Moved getBackupRouteSettingsFilePath into a route-settings adapter util
ref https://linear.app/ghost/issue/HKG-1832/implement-s3routesettingsstore

- the backup-path builder lived inside FileStore, but the upcoming
  S3RouteSettingsStore needs the same timestamped naming for its backup
  object keys; a shared util keeps one store from importing from a
  sibling store
